tests/wpt/tests/css/css-scoping/keyframes-003.html

<!doctype html>
<title>CSS Test: @keyframes applies to :host.</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<link rel="help" href="https://drafts.csswg.org/css-scoping/#shadow-names">
<div id="host"></div>
<script>
  test(function() {
    host.attachShadow({ mode: "open" }).innerHTML = `
      <style>
        @keyframes myanim {
          from { background: red; }
          to { background: green; }
        }
        :host {
          width: 100px;
          height: 100px;
          background: blue;
          animation: myanim 10s infinite;
        }
      </style>
    `;

    assert_equals(document.getElementById('host').getAnimations().length, 1);
  }, "@keyframes applies to the shadow host");
</script>
